reorganized Redux code
