+++ src/AppWrapper.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { Provider } from "react-redux";
-import { store, ConnectedComponent as Container } from "./reactReduxConnection";
+import { store } from "./reactReduxConnection/reduxStore";
+import { ConnectedComponent as Container } from "./reactReduxConnection/ConnectedComponent";
 
 export default class AppWrapper extends React.Component {
     render() {
